refactor(login): move widgets in LoginForm to LoginViewBody
